A pass at making the packages more decoupled
