+++ routes/home.go
@@ -13,6 +13,7 @@ func Home(r *mux.Router) *mux.Router {
 	r.HandleFunc("/", controllers.Homepage)
 
 	// For now...
+	r.HandleFunc("/jobs", controllers.ShowAllJobs)
 	r.HandleFunc("/companies", controllers.CompaniesIndex)
 	r.HandleFunc("/people", controllers.PeopleIndex)
 	r.HandleFunc("/people/show/{id}", controllers.NotImplemented)
